+++ demo/src/components/sections/SrStyleGuideMenu/SrStyleGuideMenu.vue
@@ -45,6 +45,8 @@
 </template>
 
 <script>
+import './SrStyleGuideMenu.scss'
+
 export default {
     name: 'SrStyleGuideMenu',
     data () {