+++ src/components/layouts/slim/SrSlimHeading/SrSlimHeading.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="d-flex w-100 justify-space-between mb-4">
+  <div class="sr-heading d-flex w-100 justify-space-between mb-4">
     <h1 class="text-h1">{{ title }}</h1>
     <div>
       <slot></slot>