+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## [new] - published date
 ### Added
-- Added SrStandardLayout component
+- SrStandardLayout component
+- Breadcrumbs from route parameters to page container
 
 ### Fixed
 - Resolved an issue where global `html` tag styles were not being applied correctly.
+- Demo page menu is now scrollable
+- Help button & breadcrumb container no longer takes up space when empty
 
 ## [1.1.0] - 2025-04-09
 ### Added